Image Compression/requirements.txt
pillow 
numpy